Add our private package repositories at install time.
When installing, if we have `npm` or `pip` present, try to set up our
private package sources. In cases where we can't, print a little
warning.

For node, this is pretty straightforward, for python it's a little
harder as `pip config` is new in `pip` 10. We try to use it if we can,
falling back to editing `pip.conf` if we need to and we know we can
without screwing other things up.

Fixes #1