Harden python package source registration
When the directory that contains the pip.conf file we need to modify
did not exist, our install script would fail, because we did not
create the intermedate directories first.

In addition, when `PIP_CONFIG_FILE` is set we should prefer that over
our default logic, as `pip` would.

Fixes #6