Freshen up the install text
This change's primary aim is to add "next steps" to the end of the
installation script, including links to our getting started guide,
examples repo, and a pointer to the 'pulumi new' command.

It also now tells you whether it is upgrading versus installing.

I could not resist adding a bit more color and personality also.