Add `pulumi-provider-build-environment` to list of ECR repos
We've been publishing the `pulumi-provider-build-environment` image to https://hub.docker.com/r/pulumi/pulumi-provider-build-environment, but our sync from Docker Hub to ECR has been failing because the `pulumi-provider-build-environment` repository hasn't existed in ECR.

We've now manually created the `pulumi-provider-build-environment` repo. This change includes it in the list so it gets the appropriate policies to allow us to publish to it.