add a new releases/dev-sdk prefix
Add a new releases/dev-sdk prefix with an expiry time of 30 days.  In
the future we'll publish all the dev releasese there, so they can be
expired after 30 days instead of being in the bucket forever and
taking up space.  We need a new prefix for this, since s3 only allows
lifecycle configuration based on prefixes, not based on a regular
expression, and we want to keep regular releases around forever.