Enable downloading Pulumi from pull releases
After this change, the install script will accept versions like `pr#16961`, and
will download and install the Pulumi binary from the corresponding pull release
artifacts.

As these artifacts are stored in GitHub CI and only accessible to authenticated
users, a GITHUB_TOKEN environment variable must be set to a valid GitHub token
with read access to the pulumi/pulumi repository. If this token is not set, the
GitHub CLI is used to obtain a token, and if that fails then an error is
presented.

```sh
./dist/install.sh --version pr#16961
=== Installing Pulumi vpr#16961 ===
+ Downloading https://api.github.com/repos/pulumi/pulumi/actions/artifacts/1808372004/zip to /tmp/pulumi.tar.gz.zip.wfe772a0CV...
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
100  179M  100  179M    0     0  14.6M      0  0:00:12  0:00:12 --:--:-- 13.5M
+ Extracting /tmp/pulumi.tar.gz.zip.wfe772a0CV to /tmp/pulumi.zip.oDmKx4Evb5...
+ Extracting to /home/friel/.pulumi/bin

=== Pulumi is now installed! 🍹 ===
+ Get started with Pulumi: https://www.pulumi.com/docs/quickstart
```

After this is released, developers can use this feature via:

```sh
curl -fsSL https://get.pulumi.com | sh -s -- --version pr#<NUMBER>
```