Add Cloudflare worker to fetch from S3 backend (#187)
Cloudflare's "Sippy" wasn't working well with + characters in a lot of our filenames.

This is basically a reimplementation of that same concept with Cloudflare Workers. If objects cannot be served from R2, the worker will reach out to our S3 backend. It will then copy the object to R2 while serving the request.

This is just the worker code. IaC using the Pulumi provider will be a separate review.

Testing:
Lots of manual poking and prodding in the dev setup. This is currently live on get-staging.pulumi.com

The big shakedown test though was grabbing a list of releases, and then taking a random sample of them and fetching the same release from both get-staging and get.pulumi.com, then diffing the two. Files were fetched in a variety of ways, using both multi-part and normal gets, and sometimes fetching it successively to test serving from R3 vs serving from R2. All files matched up perfectly.

Ref pulumi/pulumi-service#22133